backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
sentencepiece
pydantic
python-dotenv
python-multipart
gtts
SpeechRecognition
sacrebleu
nltk
sentence-transformers
numpy
pandas
aiofiles
jieba
accelerate
